+++ src/common/non_utf8/parser.rs
@@ -1,3 +1,5 @@
+use crate::no_std_compat::*;
+
 pub type ParseResult<'a, T> = Result<(ParseInput<'a>, T), ParseError>;
 pub type ParseInput<'a> = &'a [u8];
 pub type ParseError = &'static str;